pyproject.toml
[tool.poetry]
name = "base-settings-debug"
version = "0.1.0"
description = ""
authors = ["Rod Morison <rmorison@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "base_settings_debug"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pydantic = "^1.10.7"
py-moneyed = "^3.0"
rich = "^13.3.5"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
mypy = "^1.3.0"
debugpy = "^1.6.7"
pylint = "^2.17.4"
python-lsp-server = "^1.7.2"
python-lsp-black = "^1.2.1"
pyls-isort = "^0.2.2"
isort = "^5.12.0"
pylsp-mypy = "^0.6.6"
flake8 = "^6.0.0"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
pre-commit = "^3.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
plugins = ["pydantic.mypy"]
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.isort]
profile = "black"

[tool.black]
line-length = 88

[flake8]
# from https://black.readthedocs.io/en/stable/the_black_code_style/current_style.html#line-length
max-line-length = 88
select = ["C","E","F","W","B","B950"]
extend-ignore = ["E203", "E501", "W503"]
classmethod-decorators = "validator"
